Ouroboros.Consensus.Byron.Ledger

Documentation

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family Header blk ∷ Type

data UnsizedHeader

Header without a size hint

Defined in order to support backwards compatible binary encodings.

Constructors

UnsizedHeader

Fields

unsizedHeaderRaw ∷ !(ABlockOrBoundaryHdr ByteString)

unsizedHeaderSlotNo ∷ !SlotNo

unsizedHeaderHash ∷ !ByronHash

data ByronBlock

Byron block

We cache two bits of information:

We cache the slot number as this is not readily available for EBBs. Having it cached allows us to e.g. give a HasHeader instance.

We cache the hash as this is expensive to compute and we need it often.

Constructors

ByronBlock

Fields

byronBlockRaw ∷ !(ABlockOrBoundary ByteString)

byronBlockSlotNo ∷ !SlotNo

byronBlockHash ∷ !ByronHash

newtype ByronHash

Constructors

ByronHash

Fields

unByronHash ∷ HeaderHash

mkByronHash ∷ ABlockOrBoundaryHdr ByteString → ByronHash

byronHashInfo ∷ HashInfo ByronHash

mkByronBlock ∷ EpochSlots → ABlockOrBoundary ByteString → ByronBlock

annotateByronBlock ∷ EpochSlots → Block → ByronBlock

Construct Byron block from unannotated Block

This should be used only when forging blocks (not when receiving blocks over the wire).

mkByronHeader

Arguments

∷ EpochSlots

→ ABlockOrBoundaryHdr ByteString

→ SizeInBytes
Block size hint

→ Header ByronBlock

byronHeaderIsEBB ∷ Header ByronBlock → IsEBB

byronBlockIsEBB ∷ ByronBlock → IsEBB

mkUnsizedHeader ∷ EpochSlots → ABlockOrBoundaryHdr ByteString → UnsizedHeader

splitSizeHint ∷ Header ByronBlock → (UnsizedHeader, SizeInBytes)

joinSizeHint ∷ UnsizedHeader → SizeInBytes → Header ByronBlock

data ByronNodeToClientVersion

Constructors

ByronNodeToClientVersion1
No local state query protocol

ByronNodeToClientVersion2
Use local state query protocol

data ByronNodeToNodeVersion

Constructors

ByronNodeToNodeVersion1
We send headers without a size hint

ByronNodeToNodeVersion2
We send headers with a size hint

byronGenesisHash ∷ BlockConfig ByronBlock → GenesisHash

byronProtocolMagicId ∷ BlockConfig ByronBlock → ProtocolMagicId

byronProtocolMagic ∷ BlockConfig ByronBlock → ProtocolMagic

data DropEncodedSizeException

Failed to strip off the envelope from an encoded header

This indicates either a bug or disk corruption.

Constructors

DropEncodedSizeError DeserialiseFailure

DropEncodedSizeTrailingBytes ByteString

encodeByronHeaderHash ∷ HeaderHash ByronBlock → Encoding

decodeByronHeaderHash ∷ Decoder s (HeaderHash ByronBlock)

encodeByronBlockWithInfo ∷ ByronBlock → BinaryInfo Encoding

encodeByronBlock including the offset and size of the header within the resulting bytestring.

NOTE: the bytestring obtained by slicing the serialised block using the header offset and size will correspond to the header annotation, but not to the serialised header, as we add an envelope (encodeListLen + tag) around a header in encodeByronHeader. This envelope must thus still be added to the sliced bytestring before it can be deserialised using decodeByronHeader.

encodeByronBlock ∷ ByronBlock → Encoding

Encode a block

Should be backwards compatible with legacy (cardano-sl) nodes.

Implementation note: the decoder uses fromCBORABlockOrBoundary, which has inverse toCBORABlockOrBoundary. This encoder is intended to be binary compatible with toCBORABlockOrBoundary, but does not use it and instead takes advantage of the annotations (using encodePreEncoded).

decodeByronBlock ∷ EpochSlots → Decoder s (ByteString → ByronBlock)

Inverse of encodeByronBlock

encodeByronHeader ∷ SerialisationVersion ByronBlock → Header ByronBlock → Encoding

Encode a header

decodeByronHeader ∷ EpochSlots → SerialisationVersion ByronBlock → Decoder s (ByteString → Header ByronBlock)

Inverse of encodeByronHeader

encodeWrappedByronHeader ∷ SerialisationAcrossNetwork ByronBlock → Serialised (Header ByronBlock) → Encoding

Encode wrapped header

When we get a header from the chain DB, we add the missing envelope (byronAddHeaderEnvelope). This envelope is valid for version 2. If we are using version 1 on the wire, we must alter this envelope. If we are using version 2, there is nothing to do.

For either version we add a CBOR-in-CBOR wrapper.

decodeWrappedByronHeader ∷ SerialisationAcrossNetwork ByronBlock → Decoder s (Serialised (Header ByronBlock))

Decode wrapped header

See encodeWrappedByronHeader for details.

byronAddHeaderEnvelope

Arguments

∷ IsEBB

→ SizeInBytes
Block size

→ ByteString

→ ByteString

When given the raw header bytes extracted from the block, i.e., the header annotation of AHeader or ABoundaryHdr, we still need to prepend some bytes so that we can use decodeByronHeader to decode it:

fakeByronBlockSizeHint ∷ SizeInBytes

Fake size (used in compatibility mode)

mkByronContextDSIGN ∷ BlockConfig ByronBlock → VerKeyDSIGN ByronDSIGN → ContextDSIGN ByronDSIGN

Construct DSIGN required for Byron crypto

toPBftLedgerView ∷ Map → PBftLedgerView PBftByronCrypto

fromPBftLedgerView ∷ PBftLedgerView PBftByronCrypto → Map

encodeByronConsensusState ∷ ConsensusState (BlockProtocol ByronBlock) → Encoding

decodeByronConsensusState ∷ SecurityParam → Decoder s (ConsensusState (BlockProtocol ByronBlock))

verifyBlockMatchesHeader ∷ Header ByronBlock → ByronBlock → Bool

Check if a block matches its header

Note that we cannot check this for an EBB, as the EBB header doesn't store a hash of the EBB body.

verifyHeaderSignature ∷ BlockConfig ByronBlock → Header ByronBlock → Bool

Verify whether a header matches its signature.

Note that we cannot check this for an EBB, as an EBB contains no signature. This function will always return True for an EBB.

verifyHeaderIntegrity ∷ BlockConfig ByronBlock → Header ByronBlock → Bool

Verify whether a header is not corrupted.

The difference with verifyHeaderSignature is that this function also checks the integrity of the headerProtocolMagicId field, which is the only field of a regular header that is not signed.

Note that we cannot check this for an EBB, as an EBB contains no signature. This function will always return True for an EBB.

verifyBlockIntegrity ∷ BlockConfig ByronBlock → ByronBlock → Bool

Verifies whether the block is not corrupted by checking its signature and witnesses.

This function will always return True for an EBB, as we cannot check anything for an EBB.

data ByronOtherHeaderEnvelopeError

Constructors

UnexpectedEBBInSlot !SlotNo

data DelegationHistory

Delegation history

Motivation: the ledger state gives us both the current delegation state (getDelegationMap) as well any planned future changes to the delegation state (getScheduledDelegations). It does not however give us information about past delegation states. This is where the DelegationHistory comes in.

See History for more details

initByronLedgerState

Arguments

∷ Config

→ Maybe UTxO
Optionally override UTxO

→ LedgerState ByronBlock

validationErrorImpossible ∷ ∀ err a. Except err a → a

Mark computation as validation error free

Given a BlockValidationMode of NoBlockValidation, a call to applyByronBlock shouldn't fail since the ledger layer won't be performing any block validation checks. However, because applyByronBlock can fail in the event it is given a BlockValidationMode of BlockValidation, it still looks like it can fail (since its type doesn't change based on the ValidationMode) and we must still treat it as such.

encodeByronAnnTip ∷ AnnTip ByronBlock → Encoding

decodeByronAnnTip ∷ Decoder s (AnnTip ByronBlock)

encodeByronExtLedgerState ∷ ExtLedgerState ByronBlock → Encoding

encodeByronHeaderState ∷ HeaderState ByronBlock → Encoding

encodeByronLedgerState ∷ LedgerState ByronBlock → Encoding

decodeByronLedgerState ∷ Decoder s (LedgerState ByronBlock)

encodeByronQuery ∷ Query ByronBlock result → Encoding

decodeByronQuery ∷ Decoder s (Some (Query ByronBlock))

encodeByronResult ∷ Query ByronBlock result → result → Encoding

decodeByronResult ∷ Query ByronBlock result → ∀ s. Decoder s result

toMempoolPayload ∷ GenTx ByronBlock → AMempoolPayload ByteString

fromMempoolPayload ∷ AMempoolPayload ByteString → GenTx ByronBlock

byronIdTx ∷ ATxAux ByteString → TxId

byronIdDlg ∷ ACertificate ByteString → CertificateId

byronIdProp ∷ AProposal ByteString → UpId

byronIdVote ∷ AVote ByteString → VoteId

encodeByronGenTx ∷ GenTx ByronBlock → Encoding

decodeByronGenTx ∷ Decoder s (GenTx ByronBlock)

The ByteString annotation will be the canonical encoding.

While the new implementation does not care about canonical encodings, the old one does. When a generalised transaction arrives that is not in its canonical encoding (only the ATxAux of the ByronTx can be produced by nodes that are not under our control), the old implementation will reject it. Therefore, we need to reject them too. See #905.

We use the ledger to check for canonical encodings: the ledger will check whether the signed hash of the transaction (in the case of a ATxAux, the transaction witness) matches the annotated bytestring. Is therefore important that the annotated bytestring be the canonical encoding, not the original, possibly non-canonical encoding.

encodeByronGenTxId ∷ GenTxId ByronBlock → Encoding

decodeByronGenTxId ∷ Decoder s (GenTxId ByronBlock)

encodeByronApplyTxError ∷ ApplyTxErr ByronBlock → Encoding

decodeByronApplyTxError ∷ Decoder s (ApplyTxErr ByronBlock)

countByronGenTxs ∷ ByronBlock → Word64

Count all (generalized) transactions in the block

forgeByronBlock

Arguments

∷ (Monad m, HasCallStack)

⇒ TopLevelConfig ByronBlock

→ Update m (NodeState ByronBlock)

→ BlockNo
Current block number

→ TickedLedgerState ByronBlock
Current ledger

→ [GenTx ByronBlock]
Txs to add in the block

→ PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

→ m ByronBlock

forgeEBB

Arguments

∷ TopLevelConfig ByronBlock

→ SlotNo
Current slot

→ BlockNo
Current block number

→ ChainHash ByronBlock
Previous hash

→ ByronBlock

forgeRegularBlock

Arguments

∷ (Monad m, HasCallStack)

⇒ TopLevelConfig ByronBlock

→ Update m (NodeState ByronBlock)

→ BlockNo
Current block number

→ TickedLedgerState ByronBlock
Current ledger

→ [GenTx ByronBlock]
Txs to add in the block

→ PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

→ m ByronBlock

byronBlockEncodingOverhead ∷ Word32

The Byron block encoding overhead size in bytes.

This encompasses the overhead in bytes for everything that is encoded within a Byron block, excluding the actual generalized transactions (transactions, delegation certificates, update votes, and update proposals).