Ouroboros.Consensus.Byron.Ledger.Config

Block config

data family BlockConfig blk

Static configuration required to work with this type of blocks

byronEpochSlots ∷ BlockConfig ByronBlock → EpochSlots

byronGenesisHash ∷ BlockConfig ByronBlock → GenesisHash

byronProtocolMagic ∷ BlockConfig ByronBlock → ProtocolMagic

byronProtocolMagicId ∷ BlockConfig ByronBlock → ProtocolMagicId

Codec config

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

mkByronCodecConfig ∷ Config → CodecConfig ByronBlock

Storage config

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

Compact genesis config

compactGenesisConfig ∷ Config → Config

Byron's genesis config contains the AVVM balances, of which there are +14k in mainnet's genesis config. These balances are only used to create the initial ledger state, there is no reason to keep them in memory afterwards.

This function empties the gdAvvmDistr field in the genesis config. As we keep Byron's genesis config in memory (even in later eras), this can save us a bit of memory.

Orphan instances

Generic (BlockConfig ByronBlock)

Instance details

Associated Types

type Rep (BlockConfig ByronBlock) ∷ Type → Type

Methods

from ∷ BlockConfig ByronBlock → Rep (BlockConfig ByronBlock) x

to ∷ Rep (BlockConfig ByronBlock) x → BlockConfig ByronBlock

Generic (CodecConfig ByronBlock)

Instance details

Associated Types

type Rep (CodecConfig ByronBlock) ∷ Type → Type

Methods

from ∷ CodecConfig ByronBlock → Rep (CodecConfig ByronBlock) x

to ∷ Rep (CodecConfig ByronBlock) x → CodecConfig ByronBlock

Generic (StorageConfig ByronBlock)

Instance details

Associated Types

type Rep (StorageConfig ByronBlock) ∷ Type → Type

Methods

from ∷ StorageConfig ByronBlock → Rep (StorageConfig ByronBlock) x

to ∷ Rep (StorageConfig ByronBlock) x → StorageConfig ByronBlock

NoThunks (BlockConfig ByronBlock)

Instance details

Methods

noThunks ∷ Context → BlockConfig ByronBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → BlockConfig ByronBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (BlockConfig ByronBlock) → String

NoThunks (CodecConfig ByronBlock)

Instance details

Methods

noThunks ∷ Context → CodecConfig ByronBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → CodecConfig ByronBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (CodecConfig ByronBlock) → String

NoThunks (StorageConfig ByronBlock)

Instance details

Methods

noThunks ∷ Context → StorageConfig ByronBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → StorageConfig ByronBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (StorageConfig ByronBlock) → String