Ouroboros.Consensus.Byron.Ledger.DelegationHistory

Description

Delegation history

Intended for qualified import

Documentation

data DelegationHistory

Delegation history

Motivation: the ledger state gives us both the current delegation state (getDelegationMap) as well any planned future changes to the delegation state (getScheduledDelegations). It does not however give us information about past delegation states. This is where the DelegationHistory comes in.

See History for more details

empty ∷ DelegationHistory

Empty (genesis) delegation history

snapOld

Arguments

∷ BlockCount
Maximum rollback (k)

→ SlotNo
Slot number of the block that changed delegation

→ Map
Delegation state before it changed

→ DelegationHistory

→ DelegationHistory

Take a snapshot of the delegation state

find ∷ WithOrigin SlotNo → DelegationHistory → Maybe Map

Serialisation

encodeDelegationHistory ∷ DelegationHistory → Encoding

decodeDelegationHistory ∷ Decoder s DelegationHistory