Ouroboros.Consensus.Byron.Ledger.Integrity

Documentation

verifyBlockMatchesHeader ∷ Header ByronBlock → ByronBlock → Bool

Check if a block matches its header

Note that we cannot check this for an EBB, as the EBB header doesn't store a hash of the EBB body.

verifyHeaderSignature ∷ BlockConfig ByronBlock → Header ByronBlock → Bool

Verify whether a header matches its signature.

Note that we cannot check this for an EBB, as an EBB contains no signature. This function will always return True for an EBB.

verifyHeaderIntegrity ∷ BlockConfig ByronBlock → Header ByronBlock → Bool

Verify whether a header is not corrupted.

The difference with verifyHeaderSignature is that this function also checks the integrity of the headerProtocolMagicId field, which is the only field of a regular header that is not signed.

Note that we cannot check this for an EBB, as an EBB contains no signature. This function will always return True for an EBB.

verifyBlockIntegrity ∷ BlockConfig ByronBlock → ByronBlock → Bool

Verifies whether the block is not corrupted by checking its signature and witnesses.

This function will always return True for an EBB, as we cannot check anything for an EBB.