Ouroboros.Consensus.Byron.Ledger.Ledger

Description

Instances requires for consensus/ledger integration

Documentation

data ByronTransition

Transition from Byron to Shelley

Placeholder type until we properly address #2455

Constructors

ByronTransitionUnknown
Transition point not yet known

Ledger integration

initByronLedgerState

Arguments

∷ Config

→ Maybe UTxO
Optionally override UTxO

→ LedgerState ByronBlock

byronEraParams ∷ SafeBeforeEpoch → Config → EraParams

To be used for a Byron-to-X (where X is typically Shelley) chain.

byronEraParamsNeverHardForks ∷ Config → EraParams

Separate variant of byronEraParams to be used for a Byron-only chain.

Serialisation

encodeByronAnnTip ∷ AnnTip ByronBlock → Encoding

decodeByronAnnTip ∷ Decoder s (AnnTip ByronBlock)

encodeByronExtLedgerState ∷ ExtLedgerState ByronBlock → Encoding

encodeByronHeaderState ∷ HeaderState ByronBlock → Encoding

encodeByronLedgerState ∷ LedgerState ByronBlock → Encoding

decodeByronLedgerState ∷ Decoder s (LedgerState ByronBlock)

encodeByronQuery ∷ Query ByronBlock result → Encoding

decodeByronQuery ∷ Decoder s (SomeBlock Query ByronBlock)

encodeByronResult ∷ Query ByronBlock result → result → Encoding

decodeByronResult ∷ Query ByronBlock result → ∀ s. Decoder s result

Type family instances

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

Auxiliary

validationErrorImpossible ∷ ∀ err a. Except err a → a

Mark computation as validation error free

Given a BlockValidationMode of NoBlockValidation, a call to applyByronBlock shouldn't fail since the ledger layer won't be performing any block validation checks. However, because applyByronBlock can fail in the event it is given a BlockValidationMode of BlockValidation, it still looks like it can fail (since its type doesn't change based on the ValidationMode) and we must still treat it as such.

Orphan instances

HasHardForkHistory ByronBlock

Instance details

Associated Types

type HardForkIndices ByronBlock ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → Summary (HardForkIndices ByronBlock)

LedgerSupportsProtocol ByronBlock

Instance details

Methods

protocolLedgerView ∷ LedgerConfig ByronBlock → Ticked (LedgerState ByronBlock) → Ticked (LedgerView (BlockProtocol ByronBlock))

ledgerViewForecastAt ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → Forecast (LedgerView (BlockProtocol ByronBlock))

CommonProtocolParams ByronBlock

Instance details

Methods

maxHeaderSize ∷ LedgerState ByronBlock → Word32

maxTxSize ∷ LedgerState ByronBlock → Word32

UpdateLedger ByronBlock

Instance details

QueryLedger ByronBlock

Instance details

Methods

answerQuery ∷ LedgerConfig ByronBlock → Query ByronBlock result → LedgerState ByronBlock → result

Eq (LedgerState ByronBlock)

Instance details

Methods

(==) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

(/=) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

Show (LedgerState ByronBlock)

Instance details

Methods

showsPrec ∷ Int → LedgerState ByronBlock → ShowS

show ∷ LedgerState ByronBlock → String

showList ∷ [LedgerState ByronBlock] → ShowS

Generic (LedgerState ByronBlock)

Instance details

Associated Types

type Rep (LedgerState ByronBlock) ∷ Type → Type

Methods

from ∷ LedgerState ByronBlock → Rep (LedgerState ByronBlock) x

to ∷ Rep (LedgerState ByronBlock) x → LedgerState ByronBlock

Generic (Ticked (LedgerState ByronBlock))

Instance details

Associated Types

type Rep (Ticked (LedgerState ByronBlock)) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState ByronBlock) → Rep (Ticked (LedgerState ByronBlock)) x

to ∷ Rep (Ticked (LedgerState ByronBlock)) x → Ticked (LedgerState ByronBlock)

NoUnexpectedThunks (LedgerState ByronBlock)

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState ByronBlock → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState ByronBlock → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState ByronBlock) → String

NoUnexpectedThunks (Ticked (LedgerState ByronBlock))

Instance details

Methods

noUnexpectedThunks ∷ [String] → Ticked (LedgerState ByronBlock) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → Ticked (LedgerState ByronBlock) → IO ThunkInfo

showTypeOf ∷ Proxy (Ticked (LedgerState ByronBlock)) → String

GetTip (LedgerState ByronBlock)

Instance details

Methods

getTip ∷ LedgerState ByronBlock → Point (LedgerState ByronBlock)

GetTip (Ticked (LedgerState ByronBlock))

Instance details

Methods

getTip ∷ Ticked (LedgerState ByronBlock) → Point (Ticked (LedgerState ByronBlock))

IsLedger (LedgerState ByronBlock)

Instance details

Associated Types

type LedgerErr (LedgerState ByronBlock) ∷ Type

Methods

applyChainTick ∷ LedgerCfg (LedgerState ByronBlock) → SlotNo → LedgerState ByronBlock → Ticked (LedgerState ByronBlock)

SameDepIndex (Query ByronBlock)

Instance details

Methods

sameDepIndex ∷ Query ByronBlock a → Query ByronBlock b → Maybe (a :~: b)

ShowQuery (Query ByronBlock)

Instance details

Methods

showResult ∷ Query ByronBlock result → result → String

ApplyBlock (LedgerState ByronBlock) ByronBlock

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → Except (LedgerErr (LedgerState ByronBlock)) (LedgerState ByronBlock)

reapplyLedgerBlock ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → LedgerState ByronBlock

Eq (Query ByronBlock result)

Instance details

Methods

(==) ∷ Query ByronBlock result → Query ByronBlock result → Bool

(/=) ∷ Query ByronBlock result → Query ByronBlock result → Bool

Show (Query ByronBlock result)

Instance details

Methods

showsPrec ∷ Int → Query ByronBlock result → ShowS

show ∷ Query ByronBlock result → String

showList ∷ [Query ByronBlock result] → ShowS

ShowProxy (Query ByronBlock)

Instance details

Methods

showProxy ∷ Proxy (Query ByronBlock) → String