Ouroboros.Consensus.Byron.Node

Documentation

protocolInfoByron ∷ ∀ m. Monad m ⇒ Config → Maybe PBftSignatureThreshold → ProtocolVersion → SoftwareVersion → Maybe ByronLeaderCredentials → ProtocolInfo m ByronBlock

protocolClientInfoByron ∷ EpochSlots → ProtocolClientInfo ByronBlock

mkByronConfig ∷ Config → ProtocolVersion → SoftwareVersion → BlockConfig ByronBlock

newtype PBftSignatureThreshold

Signature threshold. This represents the proportion of blocks in a pbftSignatureWindow-sized window which may be signed by any single key.

Constructors

PBftSignatureThreshold

Fields

unSignatureThreshold ∷ Double

defaultPBftSignatureThreshold ∷ PBftSignatureThreshold

See chapter 4.1 of https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/byronChainSpec/latest/download-by-type/doc-pdf/blockchain-spec

byronBlockForging ∷ Monad m ⇒ ByronLeaderCredentials → BlockForging m ByronBlock

Secrets

data ByronLeaderCredentials

Constructors

ByronLeaderCredentials

Fields

blcSignKey ∷ SigningKey

blcDlgCert ∷ Certificate

blcCoreNodeId ∷ CoreNodeId

data ByronLeaderCredentialsError

mkByronLeaderCredentials ∷ Config → SigningKey → Certificate → Either ByronLeaderCredentialsError ByronLeaderCredentials

Make the ByronLeaderCredentials, with a couple sanity checks:

That the block signing key and the delegation certificate match.

That the delegation certificate does correspond to one of the genesis keys from the genesis file.

mkPBftCanBeLeader ∷ ByronLeaderCredentials → CanBeLeader (PBft PBftByronCrypto)

Orphan instances

RunNode ByronBlock

Instance details

Methods

nodeBlockFetchSize ∷ Header ByronBlock → SizeInBytes

nodeImmutableDbChunkInfo ∷ TopLevelConfig ByronBlock → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig ByronBlock → ByronBlock → Bool

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig ByronBlock → InitChainDB m ByronBlock → m ()

ConfigSupportsNode ByronBlock

Instance details

Methods

getSystemStart ∷ BlockConfig ByronBlock → SystemStart

getNetworkMagic ∷ BlockConfig ByronBlock → NetworkMagic