Ouroboros.Consensus.Byron.Node.Serialisation

Orphan instances

SerialiseNodeToNodeConstraints ByronBlock

Instance details

SerialiseNodeToClientConstraints ByronBlock

Instance details

SerialiseDiskConstraints ByronBlock

Instance details

LgrDbSerialiseConstraints ByronBlock

Instance details

ImmDbSerialiseConstraints ByronBlock

Instance details

VolDbSerialiseConstraints ByronBlock

Instance details

SerialiseNodeToNode ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s ByronBlock

SerialiseNodeToNode ByronBlock ByronHash
HeaderHash ByronBlock

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ByronHash → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s ByronHash

SerialiseNodeToClient ByronBlock ApplyMempoolPayloadErr
ApplyTxErr ByronBlock

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ApplyMempoolPayloadErr → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s ApplyMempoolPayloadErr

SerialiseNodeToClient ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s ByronBlock

SerialiseNodeToClient ByronBlock ByronHash
HeaderHash ByronBlock

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ByronHash → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s ByronHash

EncodeDisk ByronBlock ByronBlock

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → ByronBlock → Encoding

SerialiseNodeToNode ByronBlock (Serialised (Header ByronBlock))
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → Serialised (Header ByronBlock) → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s (Serialised (Header ByronBlock))

SerialiseNodeToNode ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → Serialised ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s (Serialised ByronBlock)

SerialiseNodeToNode ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → GenTx ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s (GenTx ByronBlock)

SerialiseNodeToNode ByronBlock (GenTxId ByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → GenTxId ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s (GenTxId ByronBlock)

SerialiseNodeToNode ByronBlock (Header ByronBlock)
CBOR-in-CBOR to be compatible with the wrapped (Serialised) variant. Uses nested CBOR-in-CBOR for the annotation. We can't reuse the outer CBOR-in-CBOR for the annotation because of the optional block size hint.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → Header ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → NodeToNodeVersion ByronBlock → ∀ s. Decoder s (Header ByronBlock)

SerialiseNodeToClient ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → Serialised ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s (Serialised ByronBlock)

SerialiseNodeToClient ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → GenTx ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s (GenTx ByronBlock)

SerialiseResult ByronBlock (Query ByronBlock)

Instance details

Methods

encodeResult ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → Query ByronBlock result → result → Encoding

decodeResult ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → Query ByronBlock result → ∀ s. Decoder s result

EncodeDisk ByronBlock (PBftState PBftByronCrypto)
ConsensusState (BlockProtocol ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → PBftState PBftByronCrypto → Encoding

EncodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → AnnTip ByronBlock → Encoding

EncodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → LedgerState ByronBlock → Encoding

EncodeDisk ByronBlock (Header ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → Header ByronBlock → Encoding

DecodeDisk ByronBlock (PBftState PBftByronCrypto)
ConsensusState (BlockProtocol ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (PBftState PBftByronCrypto)

DecodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (AnnTip ByronBlock)

DecodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (LedgerState ByronBlock)

SerialiseNodeToClient ByronBlock (Some (Query ByronBlock))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → Some (Query ByronBlock) → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → NodeToClientVersion ByronBlock → ∀ s. Decoder s (Some (Query ByronBlock))

DecodeDisk ByronBlock (ByteString → Header ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (ByteString → Header ByronBlock)

DecodeDisk ByronBlock (ByteString → ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (ByteString → ByronBlock)