Ouroboros.Consensus.Byron.Ledger.Block

Hash

newtype ByronHash

Constructors

ByronHash

Fields

unByronHash ∷ HeaderHash

mkByronHash ∷ ABlockOrBoundaryHdr ByteString → ByronHash

byronHashInfo ∷ HashInfo ByronHash

Block

data ByronBlock

Byron block

We cache two bits of information:

We cache the slot number as this is not readily available for EBBs. Having it cached allows us to e.g. give a HasHeader instance.

We cache the hash as this is expensive to compute and we need it often.

Constructors

ByronBlock

Fields

byronBlockRaw ∷ !(ABlockOrBoundary ByteString)

byronBlockSlotNo ∷ !SlotNo

byronBlockHash ∷ !ByronHash

mkByronBlock ∷ EpochSlots → ABlockOrBoundary ByteString → ByronBlock

annotateByronBlock ∷ EpochSlots → Block → ByronBlock

Construct Byron block from unannotated Block

This should be used only when forging blocks (not when receiving blocks over the wire).

Header

data family Header blk ∷ Type

mkByronHeader

Arguments

∷ EpochSlots

→ ABlockOrBoundaryHdr ByteString

→ SizeInBytes
Block size hint

→ Header ByronBlock

Auxiliary functions

byronHeaderIsEBB ∷ Header ByronBlock → IsEBB

byronBlockIsEBB ∷ ByronBlock → IsEBB

Low-level API

data UnsizedHeader

Header without a size hint

Defined in order to support backwards compatible binary encodings.

Constructors

UnsizedHeader

Fields

unsizedHeaderRaw ∷ !(ABlockOrBoundaryHdr ByteString)

unsizedHeaderSlotNo ∷ !SlotNo

unsizedHeaderHash ∷ !ByronHash

mkUnsizedHeader ∷ EpochSlots → ABlockOrBoundaryHdr ByteString → UnsizedHeader

splitSizeHint ∷ Header ByronBlock → (UnsizedHeader, SizeInBytes)

joinSizeHint ∷ UnsizedHeader → SizeInBytes → Header ByronBlock