Ouroboros.Consensus.Byron.Ledger.Forge

Documentation

forgeByronBlock

Arguments

∷ (MonadRandom m, HasCallStack)

⇒ TopLevelConfig ByronBlock

→ Update m (NodeState ByronBlock)

→ BlockNo
Current block number

→ TickedLedgerState ByronBlock
Current ledger

→ [GenTx ByronBlock]
Txs to add in the block

→ PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

→ m ByronBlock

forgeRegularBlock

Arguments

∷ (MonadRandom m, HasCallStack)

⇒ TopLevelConfig ByronBlock

→ Update m (NodeState ByronBlock)

→ BlockNo
Current block number

→ TickedLedgerState ByronBlock
Current ledger

→ [GenTx ByronBlock]
Txs to add in the block

→ PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

→ m ByronBlock

byronBlockEncodingOverhead ∷ Word32

The Byron block encoding overhead size in bytes.

This encompasses the overhead in bytes for everything that is encoded within a Byron block, excluding the actual generalized transactions (transactions, delegation certificates, update votes, and update proposals).

For testing purposes

forgeEBB

Arguments

∷ TopLevelConfig ByronBlock

→ SlotNo
Current slot

→ BlockNo
Current block number

→ ChainHash ByronBlock
Previous hash

→ ByronBlock