Ouroboros.Consensus.Byron.Ledger.Ledger

Description

Instances requires for consensus/ledger integration

Documentation

data ByronTransition

Information required to determine the transition from Byron to Shelley

Constructors

ByronTransitionInfo !(Map ProtocolVersion BlockNo)
Per candidate proposal, the BlockNo in which it became a candidate

The HFC needs to know when a candidate proposal becomes stable. We cannot reliably do this using SlotNo: doing so would mean that if we were to switch to a denser fork, something that was previously deemed stable is suddenly not deemed stable anymore (although in actuality it still is). We therefore must do this based on BlockNo instead, but unfortunately the Byron ledger does not record this information. Therefore, we record it here instead.

Invariant: the domain of this map should equal the set of candidate proposals.

Ledger integration

byronEraParams ∷ Config → EraParams

To be used for a Byron-to-X (where X is typically Shelley) chain.

byronEraParamsNeverHardForks ∷ Config → EraParams

Separate variant of byronEraParams to be used for a Byron-only chain.

initByronLedgerState

Arguments

∷ Config

→ Maybe UTxO
Optionally override UTxO

→ LedgerState ByronBlock

Serialisation

decodeByronAnnTip ∷ Decoder s (AnnTip ByronBlock)

decodeByronLedgerState ∷ Decoder s (LedgerState ByronBlock)

decodeByronQuery ∷ Decoder s (SomeSecond BlockQuery ByronBlock)

decodeByronResult ∷ BlockQuery ByronBlock result → ∀ s. Decoder s result

encodeByronAnnTip ∷ AnnTip ByronBlock → Encoding

encodeByronExtLedgerState ∷ ExtLedgerState ByronBlock → Encoding

encodeByronHeaderState ∷ HeaderState ByronBlock → Encoding

encodeByronLedgerState ∷ LedgerState ByronBlock → Encoding

encodeByronQuery ∷ BlockQuery ByronBlock result → Encoding

encodeByronResult ∷ BlockQuery ByronBlock result → result → Encoding

Type family instances

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family LedgerState blk

Ledger state associated with a block

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Auxiliary

validationErrorImpossible ∷ ∀ err a. Except err a → a

Mark computation as validation error free

Given a BlockValidationMode of NoBlockValidation, a call to applyByronBlock shouldn't fail since the ledger layer won't be performing any block validation checks. However, because applyByronBlock can fail in the event it is given a BlockValidationMode of BlockValidation, it still looks like it can fail (since its type doesn't change based on the ValidationMode) and we must still treat it as such.

Orphan instances

QueryLedger ByronBlock

Instance details

Methods

answerBlockQuery ∷ ExtLedgerCfg ByronBlock → BlockQuery ByronBlock result → ExtLedgerState ByronBlock → result

HasHardForkHistory ByronBlock

Instance details

Associated Types

type HardForkIndices ByronBlock ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → Summary (HardForkIndices ByronBlock)

LedgerSupportsProtocol ByronBlock

Instance details

Methods

protocolLedgerView ∷ LedgerConfig ByronBlock → Ticked (LedgerState ByronBlock) → Ticked (LedgerView (BlockProtocol ByronBlock))

ledgerViewForecastAt ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → Forecast (LedgerView (BlockProtocol ByronBlock))

CommonProtocolParams ByronBlock

Instance details

Methods

maxHeaderSize ∷ LedgerState ByronBlock → Word32

maxTxSize ∷ LedgerState ByronBlock → Word32

LedgerSupportsPeerSelection ByronBlock

Instance details

Methods

getPeers ∷ LedgerState ByronBlock → [(PoolStake, NonEmpty StakePoolRelay)]

UpdateLedger ByronBlock

Instance details

Eq (LedgerState ByronBlock)

Instance details

Methods

(==) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

(/=) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

Show (LedgerState ByronBlock)

Instance details

Methods

showsPrec ∷ Int → LedgerState ByronBlock → ShowS

show ∷ LedgerState ByronBlock → String

showList ∷ [LedgerState ByronBlock] → ShowS

Generic (LedgerState ByronBlock)

Instance details

Associated Types

type Rep (LedgerState ByronBlock) ∷ Type → Type

Methods

from ∷ LedgerState ByronBlock → Rep (LedgerState ByronBlock) x

to ∷ Rep (LedgerState ByronBlock) x → LedgerState ByronBlock

Generic (Ticked (LedgerState ByronBlock))

Instance details

Associated Types

type Rep (Ticked (LedgerState ByronBlock)) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState ByronBlock) → Rep (Ticked (LedgerState ByronBlock)) x

to ∷ Rep (Ticked (LedgerState ByronBlock)) x → Ticked (LedgerState ByronBlock)

GetTip (LedgerState ByronBlock)

Instance details

Methods

getTip ∷ LedgerState ByronBlock → Point (LedgerState ByronBlock)

GetTip (Ticked (LedgerState ByronBlock))

Instance details

Methods

getTip ∷ Ticked (LedgerState ByronBlock) → Point (Ticked (LedgerState ByronBlock))

IsLedger (LedgerState ByronBlock)

Instance details

Associated Types

type LedgerErr (LedgerState ByronBlock)

type AuxLedgerEvent (LedgerState ByronBlock)

Methods

applyChainTickLedgerResult ∷ LedgerCfg (LedgerState ByronBlock) → SlotNo → LedgerState ByronBlock → LedgerResult (LedgerState ByronBlock) (Ticked (LedgerState ByronBlock))

SameDepIndex (BlockQuery ByronBlock)

Instance details

Methods

sameDepIndex ∷ BlockQuery ByronBlock a → BlockQuery ByronBlock b → Maybe (a :~: b)

NoThunks (LedgerState ByronBlock)

Instance details

Methods

noThunks ∷ Context → LedgerState ByronBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → LedgerState ByronBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (LedgerState ByronBlock) → String

NoThunks (Ticked (LedgerState ByronBlock))

Instance details

Methods

noThunks ∷ Context → Ticked (LedgerState ByronBlock) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Ticked (LedgerState ByronBlock) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Ticked (LedgerState ByronBlock)) → String

ShowQuery (BlockQuery ByronBlock)

Instance details

Methods

showResult ∷ BlockQuery ByronBlock result → result → String

ApplyBlock (LedgerState ByronBlock) ByronBlock

Instance details

Methods

applyBlockLedgerResult ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → Except (LedgerErr (LedgerState ByronBlock)) (LedgerResult (LedgerState ByronBlock) (LedgerState ByronBlock))

reapplyBlockLedgerResult ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → LedgerResult (LedgerState ByronBlock) (LedgerState ByronBlock)

Eq (BlockQuery ByronBlock result)

Instance details

Methods

(==) ∷ BlockQuery ByronBlock result → BlockQuery ByronBlock result → Bool

(/=) ∷ BlockQuery ByronBlock result → BlockQuery ByronBlock result → Bool

Show (BlockQuery ByronBlock result)

Instance details

Methods

showsPrec ∷ Int → BlockQuery ByronBlock result → ShowS

show ∷ BlockQuery ByronBlock result → String

showList ∷ [BlockQuery ByronBlock result] → ShowS

ShowProxy (BlockQuery ByronBlock)

Instance details

Methods

showProxy ∷ Proxy (BlockQuery ByronBlock) → String