Ouroboros.Consensus.Byron.Ledger.Ledger

Description

Instances requires for consensus/ledger integration

Ledger integration

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

initByronLedgerState

Arguments

∷ Config

→ Maybe UTxO
Optionally override UTxO

→ LedgerState ByronBlock

Serialisation

encodeByronAnnTip ∷ AnnTip ByronBlock → Encoding

decodeByronAnnTip ∷ Decoder s (AnnTip ByronBlock)

encodeByronExtLedgerState ∷ ExtLedgerState ByronBlock → Encoding

encodeByronHeaderState ∷ HeaderState ByronBlock → Encoding

encodeByronLedgerState ∷ LedgerState ByronBlock → Encoding

decodeByronLedgerState ∷ Decoder s (LedgerState ByronBlock)

encodeByronQuery ∷ Query ByronBlock result → Encoding

decodeByronQuery ∷ Decoder s (Some (Query ByronBlock))

encodeByronResult ∷ Query ByronBlock result → result → Encoding

decodeByronResult ∷ Query ByronBlock result → ∀ s. Decoder s result

Auxiliary

validationErrorImpossible ∷ ∀ err a. Except err a → a

Mark computation as validation error free

Given a BlockValidationMode of NoBlockValidation, a call to applyByronBlock shouldn't fail since the ledger layer won't be performing any block validation checks. However, because applyByronBlock can fail in the event it is given a BlockValidationMode of BlockValidation, it still looks like it can fail (since its type doesn't change based on the ValidationMode) and we must still treat it as such.

Orphan instances

LedgerSupportsProtocol ByronBlock

Instance details

Methods

protocolLedgerView ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → LedgerView (BlockProtocol ByronBlock)

ledgerViewForecastAt ∷ LedgerConfig ByronBlock → LedgerState ByronBlock → WithOrigin SlotNo → Maybe (Forecast (LedgerView (BlockProtocol ByronBlock)))

HasHardForkHistory ByronBlock

Instance details

Associated Types

type HardForkIndices ByronBlock ∷ [Type]

Methods

hardForkSummary ∷ SystemStart → LedgerConfig ByronBlock → LedgerState ByronBlock → Summary (HardForkIndices ByronBlock)

UpdateLedger ByronBlock

Instance details

QueryLedger ByronBlock

Instance details

Associated Types

data Query ByronBlock a ∷ Type

Methods

answerQuery ∷ LedgerConfig ByronBlock → Query ByronBlock result → LedgerState ByronBlock → result

eqQuery ∷ Query ByronBlock result1 → Query ByronBlock result2 → Maybe (result1 :~: result2)

Eq (LedgerState ByronBlock)

Instance details

Methods

(==) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

(/=) ∷ LedgerState ByronBlock → LedgerState ByronBlock → Bool

Show (LedgerState ByronBlock)

Instance details

Methods

showsPrec ∷ Int → LedgerState ByronBlock → ShowS

show ∷ LedgerState ByronBlock → String

showList ∷ [LedgerState ByronBlock] → ShowS

Generic (LedgerState ByronBlock)

Instance details

Associated Types

type Rep (LedgerState ByronBlock) ∷ Type → Type

Methods

from ∷ LedgerState ByronBlock → Rep (LedgerState ByronBlock) x

to ∷ Rep (LedgerState ByronBlock) x → LedgerState ByronBlock

NoUnexpectedThunks (LedgerState ByronBlock)

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState ByronBlock → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState ByronBlock → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState ByronBlock) → String

IsLedger (LedgerState ByronBlock)

Instance details

Associated Types

type LedgerErr (LedgerState ByronBlock) ∷ Type

Methods

applyChainTick ∷ LedgerCfg (LedgerState ByronBlock) → SlotNo → LedgerState ByronBlock → Ticked (LedgerState ByronBlock)

ShowQuery (Query ByronBlock)

Instance details

Methods

showResult ∷ Query ByronBlock result → result → String

ApplyBlock (LedgerState ByronBlock) ByronBlock

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → Except (LedgerErr (LedgerState ByronBlock)) (LedgerState ByronBlock)

reapplyLedgerBlock ∷ LedgerCfg (LedgerState ByronBlock) → ByronBlock → Ticked (LedgerState ByronBlock) → LedgerState ByronBlock

ledgerTipPoint ∷ LedgerState ByronBlock → Point ByronBlock

Eq (Query ByronBlock result)

Instance details

Methods

(==) ∷ Query ByronBlock result → Query ByronBlock result → Bool

(/=) ∷ Query ByronBlock result → Query ByronBlock result → Bool

Show (Query ByronBlock result)

Instance details

Methods

showsPrec ∷ Int → Query ByronBlock result → ShowS

show ∷ Query ByronBlock result → String

showList ∷ [Query ByronBlock result] → ShowS