Ouroboros.Consensus.Byron.Ledger.Mempool

Description

Byron mempool integration

Mempool integration

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx

A generalized transaction, GenTx, identifier.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

Transaction IDs

byronIdDlg ∷ ACertificate ByteString → CertificateId

byronIdProp ∷ AProposal ByteString → UpId

byronIdTx ∷ ATxAux ByteString → TxId

byronIdVote ∷ AVote ByteString → VoteId

Serialisation

decodeByronApplyTxError ∷ Decoder s (ApplyTxErr ByronBlock)

decodeByronGenTx ∷ Decoder s (GenTx ByronBlock)

The ByteString annotation will be the canonical encoding.

While the new implementation does not care about canonical encodings, the old one does. When a generalised transaction arrives that is not in its canonical encoding (only the ATxAux of the ByronTx can be produced by nodes that are not under our control), the old implementation will reject it. Therefore, we need to reject them too. See #905.

We use the ledger to check for canonical encodings: the ledger will check whether the signed hash of the transaction (in the case of a ATxAux, the transaction witness) matches the annotated bytestring. Is therefore important that the annotated bytestring be the canonical encoding, not the original, possibly non-canonical encoding.

decodeByronGenTxId ∷ Decoder s (GenTxId ByronBlock)

encodeByronApplyTxError ∷ ApplyTxErr ByronBlock → Encoding

encodeByronGenTx ∷ GenTx ByronBlock → Encoding

encodeByronGenTxId ∷ GenTxId ByronBlock → Encoding

Low-level API (primarily for testing)

fromMempoolPayload ∷ AMempoolPayload ByteString → GenTx ByronBlock

toMempoolPayload ∷ GenTx ByronBlock → AMempoolPayload ByteString

Auxiliary functions

countByronGenTxs ∷ ByronBlock → Word64

Count all (generalized) transactions in the block

Orphan instances

TxLimits ByronBlock

Instance details

Associated Types

type TxMeasure ByronBlock

Methods

txMeasure ∷ Validated (GenTx ByronBlock) → TxMeasure ByronBlock

txsBlockCapacity ∷ Ticked (LedgerState ByronBlock) → TxMeasure ByronBlock

LedgerSupportsMempool ByronBlock

Instance details

Methods

txInvariant ∷ GenTx ByronBlock → Bool

applyTx ∷ LedgerConfig ByronBlock → WhetherToIntervene → SlotNo → GenTx ByronBlock → TickedLedgerState ByronBlock → Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock, Validated (GenTx ByronBlock))

reapplyTx ∷ LedgerConfig ByronBlock → SlotNo → Validated (GenTx ByronBlock) → TickedLedgerState ByronBlock → Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock)

txsMaxBytes ∷ TickedLedgerState ByronBlock → Word32

txInBlockSize ∷ GenTx ByronBlock → Word32

txForgetValidated ∷ Validated (GenTx ByronBlock) → GenTx ByronBlock

HasTxs ByronBlock

Instance details

Methods

extractTxs ∷ ByronBlock → [GenTx ByronBlock]

ShowProxy (GenTx ByronBlock)

Instance details

Methods

showProxy ∷ Proxy (GenTx ByronBlock) → String

ShowProxy (TxId (GenTx ByronBlock))

Instance details

Methods

showProxy ∷ Proxy (TxId (GenTx ByronBlock)) → String

Eq (GenTx ByronBlock)

Instance details

Methods

(==) ∷ GenTx ByronBlock → GenTx ByronBlock → Bool

(/=) ∷ GenTx ByronBlock → GenTx ByronBlock → Bool

Eq (TxId (GenTx ByronBlock))

Instance details

Methods

(==) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(/=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

Eq (Validated (GenTx ByronBlock))

Instance details

Methods

(==) ∷ Validated (GenTx ByronBlock) → Validated (GenTx ByronBlock) → Bool

(/=) ∷ Validated (GenTx ByronBlock) → Validated (GenTx ByronBlock) → Bool

Ord (TxId (GenTx ByronBlock))

Instance details

Methods

compare ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Ordering

(<) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(<=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(>) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(>=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

max ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock)

min ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock)

Show (GenTx ByronBlock)

Instance details

Methods

showsPrec ∷ Int → GenTx ByronBlock → ShowS

show ∷ GenTx ByronBlock → String

showList ∷ [GenTx ByronBlock] → ShowS

Show (GenTxId ByronBlock)

Instance details

Methods

showsPrec ∷ Int → GenTxId ByronBlock → ShowS

show ∷ GenTxId ByronBlock → String

showList ∷ [GenTxId ByronBlock] → ShowS

Show (Validated (GenTx ByronBlock))

Instance details

Methods

showsPrec ∷ Int → Validated (GenTx ByronBlock) → ShowS

show ∷ Validated (GenTx ByronBlock) → String

showList ∷ [Validated (GenTx ByronBlock)] → ShowS

Generic (GenTx ByronBlock)

Instance details

Associated Types

type Rep (GenTx ByronBlock) ∷ Type → Type

Methods

from ∷ GenTx ByronBlock → Rep (GenTx ByronBlock) x

to ∷ Rep (GenTx ByronBlock) x → GenTx ByronBlock

Generic (Validated (GenTx ByronBlock))

Instance details

Associated Types

type Rep (Validated (GenTx ByronBlock)) ∷ Type → Type

Methods

from ∷ Validated (GenTx ByronBlock) → Rep (Validated (GenTx ByronBlock)) x

to ∷ Rep (Validated (GenTx ByronBlock)) x → Validated (GenTx ByronBlock)

HasTxId (GenTx ByronBlock)

Instance details

Methods

txId ∷ GenTx ByronBlock → TxId (GenTx ByronBlock)

Condense (GenTx ByronBlock)

Instance details

Methods

condense ∷ GenTx ByronBlock → String

Condense (GenTxId ByronBlock)

Instance details

Methods

condense ∷ GenTxId ByronBlock → String

NoThunks (GenTx ByronBlock)

Instance details

Methods

noThunks ∷ Context → GenTx ByronBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → GenTx ByronBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (GenTx ByronBlock) → String

NoThunks (TxId (GenTx ByronBlock))

Instance details

Methods

noThunks ∷ Context → TxId (GenTx ByronBlock) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → TxId (GenTx ByronBlock) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (TxId (GenTx ByronBlock)) → String

NoThunks (Validated (GenTx ByronBlock))

Instance details

Methods

noThunks ∷ Context → Validated (GenTx ByronBlock) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Validated (GenTx ByronBlock) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Validated (GenTx ByronBlock)) → String

ShowProxy ApplyMempoolPayloadErr

Instance details

Methods

showProxy ∷ Proxy ApplyMempoolPayloadErr → String