Ouroboros.Consensus.Byron.Ledger.Mempool

Description

Byron mempool integration

Mempool integration

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

Transaction IDs

byronIdTx ∷ ATxAux ByteString → TxId

byronIdDlg ∷ ACertificate ByteString → CertificateId

byronIdProp ∷ AProposal ByteString → UpId

byronIdVote ∷ AVote ByteString → VoteId

Serialisation

encodeByronGenTx ∷ GenTx ByronBlock → Encoding

decodeByronGenTx ∷ Decoder s (GenTx ByronBlock)

The ByteString annotation will be the canonical encoding.

While the new implementation does not care about canonical encodings, the old one does. When a generalised transaction arrives that is not in its canonical encoding (only the ATxAux of the ByronTx can be produced by nodes that are not under our control), the old implementation will reject it. Therefore, we need to reject them too. See #905.

We use the ledger to check for canonical encodings: the ledger will check whether the signed hash of the transaction (in the case of a ATxAux, the transaction witness) matches the annotated bytestring. Is therefore important that the annotated bytestring be the canonical encoding, not the original, possibly non-canonical encoding.

encodeByronGenTxId ∷ GenTxId ByronBlock → Encoding

decodeByronGenTxId ∷ Decoder s (GenTxId ByronBlock)

encodeByronApplyTxError ∷ ApplyTxErr ByronBlock → Encoding

decodeByronApplyTxError ∷ Decoder s (ApplyTxErr ByronBlock)

Low-level API (primarily for testing)

toMempoolPayload ∷ GenTx ByronBlock → AMempoolPayload ByteString

fromMempoolPayload ∷ AMempoolPayload ByteString → GenTx ByronBlock

Auxiliary functions

countByronGenTxs ∷ ByronBlock → Word64

Count all (generalized) transactions in the block

Orphan instances

ApplyTx ByronBlock

Instance details

Associated Types

data GenTx ByronBlock ∷ Type

type ApplyTxErr ByronBlock ∷ Type

Methods

txInvariant ∷ GenTx ByronBlock → Bool

applyTx ∷ LedgerConfig ByronBlock → GenTx ByronBlock → TickedLedgerState ByronBlock → Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock)

reapplyTx ∷ LedgerConfig ByronBlock → GenTx ByronBlock → TickedLedgerState ByronBlock → Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock)

HasTxs ByronBlock

Instance details

Methods

extractTxs ∷ ByronBlock → [GenTx ByronBlock]

Eq (GenTx ByronBlock)

Instance details

Methods

(==) ∷ GenTx ByronBlock → GenTx ByronBlock → Bool

(/=) ∷ GenTx ByronBlock → GenTx ByronBlock → Bool

Eq (TxId (GenTx ByronBlock))

Instance details

Methods

(==) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(/=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

Ord (TxId (GenTx ByronBlock))

Instance details

Methods

compare ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Ordering

(<) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(<=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(>) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

(>=) ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → Bool

max ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock)

min ∷ TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock) → TxId (GenTx ByronBlock)

Show (GenTx ByronBlock)

Instance details

Methods

showsPrec ∷ Int → GenTx ByronBlock → ShowS

show ∷ GenTx ByronBlock → String

showList ∷ [GenTx ByronBlock] → ShowS

Show (GenTxId ByronBlock)

Instance details

Methods

showsPrec ∷ Int → GenTxId ByronBlock → ShowS

show ∷ GenTxId ByronBlock → String

showList ∷ [GenTxId ByronBlock] → ShowS

Generic (GenTx ByronBlock)

Instance details

Associated Types

type Rep (GenTx ByronBlock) ∷ Type → Type

Methods

from ∷ GenTx ByronBlock → Rep (GenTx ByronBlock) x

to ∷ Rep (GenTx ByronBlock) x → GenTx ByronBlock

NoUnexpectedThunks (GenTx ByronBlock)

Instance details

Methods

noUnexpectedThunks ∷ [String] → GenTx ByronBlock → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → GenTx ByronBlock → IO ThunkInfo

showTypeOf ∷ Proxy (GenTx ByronBlock) → String

NoUnexpectedThunks (TxId (GenTx ByronBlock))

Instance details

Methods

noUnexpectedThunks ∷ [String] → TxId (GenTx ByronBlock) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → TxId (GenTx ByronBlock) → IO ThunkInfo

showTypeOf ∷ Proxy (TxId (GenTx ByronBlock)) → String

HasTxId (GenTx ByronBlock)

Instance details

Associated Types

data TxId (GenTx ByronBlock) ∷ Type

Methods

txId ∷ GenTx ByronBlock → TxId (GenTx ByronBlock)

Condense (GenTx ByronBlock)

Instance details

Methods

condense ∷ GenTx ByronBlock → String

Condense (GenTxId ByronBlock)

Instance details

Methods

condense ∷ GenTxId ByronBlock → String