Ouroboros.Consensus.Byron.Ledger.Serialisation

Serialisation

encodeByronBlockWithInfo ∷ ByronBlock → BinaryInfo Encoding

encodeByronBlock including the offset and size of the header within the resulting bytestring.

NOTE: the bytestring obtained by slicing the serialised block using the header offset and size will correspond to the header annotation, but not to the serialised header, as we add an envelope (encodeListLen + tag) around a header in encodeByronHeader. This envelope must thus still be added to the sliced bytestring before it can be deserialised using decodeByronHeader.

encodeByronBlock ∷ ByronBlock → Encoding

Encode a block

Should be backwards compatible with legacy (cardano-sl) nodes.

Implementation note: the decoder uses fromCBORABlockOrBoundary, which has inverse toCBORABlockOrBoundary. This encoder is intended to be binary compatible with toCBORABlockOrBoundary, but does not use it and instead takes advantage of the annotations (using encodePreEncoded).

decodeByronBlock ∷ EpochSlots → Decoder s (ByteString → ByronBlock)

Inverse of encodeByronBlock

encodeByronHeader ∷ SerialisationVersion ByronNetworkProtocolVersion → Header ByronBlock → Encoding

Encode a header

decodeByronHeader ∷ EpochSlots → SerialisationVersion ByronNetworkProtocolVersion → Decoder s (ByteString → Header ByronBlock)

Inverse of encodeByronHeader

encodeWrappedByronHeader ∷ ByronNetworkProtocolVersion → Serialised (Header ByronBlock) → Encoding

Encode wrapped header

When we get a header from the chain DB, we add the missing envelope (byronAddHeaderEnvelope). This envelope is valid for version 2. If we are using version 1 on the wire, we must alter this envelope. If we are using version 2, there is nothing to do.

For either version we add a CBOR-in-CBOR wrapper.

decodeWrappedByronHeader ∷ ByronNetworkProtocolVersion → Decoder s (Serialised (Header ByronBlock))

Decode wrapped header

See encodeWrappedByronHeader for details.

encodeByronHeaderHash ∷ HeaderHash ByronBlock → Encoding

decodeByronHeaderHash ∷ Decoder s (HeaderHash ByronBlock)

Support for on-disk format

byronAddHeaderEnvelope

Arguments

∷ IsEBB

→ SizeInBytes
Block size

→ ByteString

→ ByteString

When given the raw header bytes extracted from the block, i.e., the header annotation of AHeader or ABoundaryHdr, we still need to prepend some bytes so that we can use decodeByronHeader to decode it:

Exceptions

data DropEncodedSizeException

Failed to strip off the envelope from an encoded header

This indicates either a bug or disk corruption.

Constructors

DropEncodedSizeError DeserialiseFailure

DropEncodedSizeTrailingBytes ByteString

Low-level API

fakeByronBlockSizeHint ∷ SizeInBytes

Fake size (used in compatibility mode)

Orphan instances

Serialise ByronHash

Instance details

Methods

encode ∷ ByronHash → Encoding

decode ∷ Decoder s ByronHash

encodeList ∷ [ByronHash] → Encoding

decodeList ∷ Decoder s [ByronHash]