Ouroboros.Consensus.Byron.Node

Documentation

protocolInfoByron ∷ Config → Maybe PBftSignatureThreshold → ProtocolVersion → SoftwareVersion → Maybe PBftLeaderCredentials → ProtocolInfo ByronBlock

mkByronConfig ∷ Config → ProtocolVersion → SoftwareVersion → BlockConfig ByronBlock

newtype PBftSignatureThreshold

Signature threshold. This represents the proportion of blocks in a pbftSignatureWindow-sized window which may be signed by any single key.

Constructors

PBftSignatureThreshold

Fields

unSignatureThreshold ∷ Double

defaultPBftSignatureThreshold ∷ PBftSignatureThreshold

See chapter 4.1 of https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/byronChainSpec/latest/download-by-type/doc-pdf/blockchain-spec

Secrets

data PBftLeaderCredentials

data PBftLeaderCredentialsError

mkPBftLeaderCredentials ∷ Config → SigningKey → Certificate → Either PBftLeaderCredentialsError PBftLeaderCredentials

Make the PBftLeaderCredentials, with a couple sanity checks:

That the block signing key and the delegation certificate match.

That the delegation certificate does correspond to one of the genesis keys from the genesis file.

pbftLeaderOrNot ∷ PBftLeaderCredentials → PBftIsLeader PBftByronCrypto

For testing

plcCoreNodeId ∷ PBftLeaderCredentials → CoreNodeId

Exported for the benefit of ByronDual

extractEpochSlots ∷ BlockConfig ByronBlock → EpochSlots

Orphan instances

RunNode ByronBlock

Instance details

Methods

nodeForgeBlock ∷ MonadRandom m ⇒ TopLevelConfig ByronBlock → Update m (NodeState ByronBlock) → BlockNo → TickedLedgerState ByronBlock → [GenTx ByronBlock] → IsLeader (BlockProtocol ByronBlock) → m ByronBlock

nodeBlockMatchesHeader ∷ Header ByronBlock → ByronBlock → Bool

nodeBlockFetchSize ∷ Header ByronBlock → SizeInBytes

nodeIsEBB ∷ Header ByronBlock → Maybe EpochNo

nodeImmDbChunkInfo ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → ChunkInfo

nodeStartTime ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → SystemStart

nodeNetworkMagic ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → NetworkMagic

nodeProtocolMagicId ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → ProtocolMagicId

nodeHashInfo ∷ Proxy ByronBlock → HashInfo (HeaderHash ByronBlock)

nodeMaxBlockSize ∷ LedgerState ByronBlock → Word32

nodeBlockEncodingOverhead ∷ LedgerState ByronBlock → Word32

nodeCheckIntegrity ∷ TopLevelConfig ByronBlock → ByronBlock → Bool

nodeAddHeaderEnvelope ∷ Proxy ByronBlock → IsEBB → SizeInBytes → ByteString → ByteString

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig ByronBlock → ChainDB m ByronBlock → m ()

nodeExceptionIsFatal ∷ Proxy ByronBlock → SomeException → Maybe ExitReason

nodeEncodeBlockWithInfo ∷ BlockConfig ByronBlock → ByronBlock → BinaryInfo Encoding

nodeEncodeBlock ∷ BlockConfig ByronBlock → ByronBlock → Encoding

nodeEncodeHeader ∷ BlockConfig ByronBlock → SerialisationVersion ByronBlock → Header ByronBlock → Encoding

nodeEncodeWrappedHeader ∷ BlockConfig ByronBlock → SerialisationAcrossNetwork ByronBlock → Serialised (Header ByronBlock) → Encoding

nodeEncodeGenTx ∷ GenTx ByronBlock → Encoding

nodeEncodeGenTxId ∷ GenTxId ByronBlock → Encoding

nodeEncodeHeaderHash ∷ Proxy ByronBlock → HeaderHash ByronBlock → Encoding

nodeEncodeLedgerState ∷ LedgerState ByronBlock → Encoding

nodeEncodeConsensusState ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → ConsensusState (BlockProtocol ByronBlock) → Encoding

nodeEncodeApplyTxError ∷ Proxy ByronBlock → ApplyTxErr ByronBlock → Encoding

nodeEncodeTipInfo ∷ Proxy ByronBlock → TipInfo ByronBlock → Encoding

nodeEncodeQuery ∷ Query ByronBlock result → Encoding

nodeEncodeResult ∷ Query ByronBlock result → result → Encoding

nodeDecodeHeader ∷ BlockConfig ByronBlock → SerialisationVersion ByronBlock → Decoder s (ByteString → Header ByronBlock)

nodeDecodeWrappedHeader ∷ BlockConfig ByronBlock → SerialisationAcrossNetwork ByronBlock → Decoder s (Serialised (Header ByronBlock))

nodeDecodeBlock ∷ BlockConfig ByronBlock → Decoder s (ByteString → ByronBlock)

nodeDecodeGenTx ∷ Decoder s (GenTx ByronBlock)

nodeDecodeGenTxId ∷ Decoder s (GenTxId ByronBlock)

nodeDecodeHeaderHash ∷ Proxy ByronBlock → Decoder s (HeaderHash ByronBlock)

nodeDecodeLedgerState ∷ Decoder s (LedgerState ByronBlock)

nodeDecodeConsensusState ∷ Proxy ByronBlock → TopLevelConfig ByronBlock → Decoder s (ConsensusState (BlockProtocol ByronBlock))

nodeDecodeApplyTxError ∷ Proxy ByronBlock → Decoder s (ApplyTxErr ByronBlock)

nodeDecodeTipInfo ∷ Proxy ByronBlock → Decoder s (TipInfo ByronBlock)

nodeDecodeQuery ∷ Decoder s (Some (Query ByronBlock))

nodeDecodeResult ∷ Query ByronBlock result → ∀ s. Decoder s result