Ouroboros.Consensus.Byron.Node.Serialisation

Orphan instances

SerialiseNodeToNodeConstraints ByronBlock

Instance details

SerialiseNodeToClientConstraints ByronBlock

Instance details

SerialiseDiskConstraints ByronBlock

Instance details

LgrDbSerialiseConstraints ByronBlock

Instance details

ImmDbSerialiseConstraints ByronBlock

Instance details

VolDbSerialiseConstraints ByronBlock

Instance details

EncodeDisk ByronBlock ByronBlock

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → ByronBlock → Encoding

ReconstructNestedCtxt Header ByronBlock

Instance details

Methods

reconstructPrefixLen ∷ proxy (Header ByronBlock) → Word8

reconstructNestedCtxt ∷ proxy (Header ByronBlock) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) ByronBlock

SerialiseNodeToNode ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s ByronBlock

SerialiseNodeToNode ByronBlock ByronHash
HeaderHash ByronBlock

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ByronHash → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s ByronHash

SerialiseNodeToClient ByronBlock ApplyMempoolPayloadErr
ApplyTxErr ByronBlock

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ApplyMempoolPayloadErr → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s ApplyMempoolPayloadErr

SerialiseNodeToClient ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s ByronBlock

SerialiseNodeToClient ByronBlock ByronHash
HeaderHash ByronBlock

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ByronHash → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s ByronHash

EncodeDisk ByronBlock (PBftState PBftByronCrypto)
ConsensusState (BlockProtocol ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → PBftState PBftByronCrypto → Encoding

EncodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → AnnTip ByronBlock → Encoding

EncodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

encodeDisk ∷ CodecConfig ByronBlock → LedgerState ByronBlock → Encoding

DecodeDisk ByronBlock (PBftState PBftByronCrypto)
ConsensusState (BlockProtocol ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (PBftState PBftByronCrypto)

DecodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (AnnTip ByronBlock)

DecodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (LedgerState ByronBlock)

SerialiseNodeToNode ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → Serialised ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s (Serialised ByronBlock)

SerialiseNodeToNode ByronBlock (SerialisedHeader ByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → SerialisedHeader ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s (SerialisedHeader ByronBlock)

SerialiseNodeToNode ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → GenTx ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s (GenTx ByronBlock)

SerialiseNodeToNode ByronBlock (GenTxId ByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → GenTxId ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s (GenTxId ByronBlock)

SerialiseNodeToNode ByronBlock (Header ByronBlock)

Instance details

Methods

encodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → Header ByronBlock → Encoding

decodeNodeToNode ∷ CodecConfig ByronBlock → BlockNodeToNodeVersion ByronBlock → ∀ s. Decoder s (Header ByronBlock)

SerialiseNodeToClient ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → Serialised ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s (Serialised ByronBlock)

SerialiseNodeToClient ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → GenTx ByronBlock → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s (GenTx ByronBlock)

SerialiseResult ByronBlock (Query ByronBlock)

Instance details

Methods

encodeResult ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → Query ByronBlock result → result → Encoding

decodeResult ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → Query ByronBlock result → ∀ s. Decoder s result

DecodeDisk ByronBlock (ByteString → ByronBlock)

Instance details

Methods

decodeDisk ∷ CodecConfig ByronBlock → ∀ s. Decoder s (ByteString → ByronBlock)

SerialiseNodeToClient ByronBlock (Some (Query ByronBlock))

Instance details

Methods

encodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → Some (Query ByronBlock) → Encoding

decodeNodeToClient ∷ CodecConfig ByronBlock → BlockNodeToClientVersion ByronBlock → ∀ s. Decoder s (Some (Query ByronBlock))

EncodeDiskDepIx (NestedCtxt Header) ByronBlock

Instance details

Methods

encodeDiskDepIx ∷ CodecConfig ByronBlock → SomeBlock (NestedCtxt Header) ByronBlock → Encoding

EncodeDiskDep (NestedCtxt Header) ByronBlock

Instance details

Methods

encodeDiskDep ∷ CodecConfig ByronBlock → NestedCtxt Header ByronBlock a → a → Encoding

DecodeDiskDepIx (NestedCtxt Header) ByronBlock

Instance details

Methods

decodeDiskDepIx ∷ CodecConfig ByronBlock → Decoder s (SomeBlock (NestedCtxt Header) ByronBlock)

DecodeDiskDep (NestedCtxt Header) ByronBlock

Instance details

Methods

decodeDiskDep ∷ CodecConfig ByronBlock → NestedCtxt Header ByronBlock a → ∀ s. Decoder s (ByteString → a)